requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
cryptography==44.0.0
    # via -r requirements.in
django==5.1.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-cors-headers==4.6.0
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular[sidecar]==0.28.0
    # via -r requirements.in
drf-spectacular-sidecar==2024.12.1
    # via drf-spectacular
envparse==0.2.0
    # via -r requirements.in
factory-boy==3.3.1
    # via -r requirements.in
faker==33.1.0
    # via factory-boy
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
packaging==24.2
    # via build
pillow==11.0.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via faker
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r requirements.in
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via faker
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
